Remove util/deployapi. Update init_cluster.sh. Rewrite Dockerfile
Update brew packages.
Update Cargo pkgs.

+++ scripts/init_cluster.sh
@@ -1,9 +1,23 @@
 #!/bin/bash
 set -euo pipefail
 
-if [[ -z "$GCP_PROJECT_ID" ]]; then
-    echo '$GCP_PROJECT_ID not set'
-    echo 'Please rerun with GCP_PROJECT_ID=my-project ./scripts/init_cluster.sh'
+# if the number of arguments isn't 1
+if [[ "$#" -ne 1 ]]; then
+    echo 'Usage: ./init_cluster.sh <GCP-PROJECT-ID>'
+    echo ''
+    echo 'Examples:'
+    echo '  ./scripts/init_cluster.sh studybeast-dev-ryan-boehning'
+    exit 1
+fi
+
+GCP_PROJECT_ID="$1"
+if [[ "$GCP_PROJECT_ID" == "studybeast-prod" ]] ||
+    [[ "$GCP_PROJECT_ID" == "studybeast-staging" ]] ||
+    [[ "$GCP_PROJECT_ID" =~ studybeast\-dev\-[a-z-]+ ]]; then
+    : # project id is valid
+else
+    echo "Invalid project id '$GCP_PROJECT_ID'"
+    echo 'Project id must match one of these: studybeast-prod, studybeast-staging, studybeast-dev-[a-z-]+'
     exit 1
 fi
 
@@ -24,12 +38,14 @@ gcloud services enable \
 #   .[] | {name,network} means "create a list of objects with just the name and network keys"
 #   select(.network | contains("foo")) means "if the network contains the foo string"
 #   .name means "only output the firewall rule name"
-gcloud compute firewall-rules list --format=json |
-    jq -r '.[] | {name,network} | select(.network | contains("global/networks/default")) | .name' |
+#   || true make the command succeed if the default network has already been deleted (idempotence)
+RULES=$(gcloud compute firewall-rules list --format=json | jq -r '.[] | {name,network} | select(.network | endswith("global/networks/default")) | .name')
+if [[ -z "$RULES" ]]; then
     xargs -I {} gcloud compute firewall-rules delete --quiet {}
+fi
 
 # Delete the default VPC
-gcloud compute networks delete --quiet default
-
-# TODO: create GCR registry
-# TODO: docker build prod API image and push to GCR
+VPCS=$(gcloud compute network list --format=json | jq -r '.name')
+if [[ -z "$VPCS" ]]; then
+    gcloud compute networks delete --quiet default
+fi
